Move non-device-specific methods to neurostack class
